Add currently existing NGINX conf file
This is the configuration file that nginx is currently using on the server
side. This file will be changed shortly to accomodate both React and Express
running together on the same server once I figure out how to accomodate that.
Deployment scripts also to come. Note to future me: this file actually resides
in `/etc/nginx/sites-available/tutorial`.