Add tentative backend deployment script
This assumes that we don't need to restart any servers since Nodemon should be
keeping watch for us when files are changed. Going to set this up soon so that
all of the Express code I've written is running on the deployment environment
and I can actually test and prototype things for real rather than with toy
Python scripts.