Comment out most of the routes for easy testing purposes
It's kind of a pain to test things with all of the routes enabled, since the
file essentially expects all of the functions to exist in some form. Rather
than write out a couple dozen dummy function bodies, I'm going to comment
almost all of them out for now and create function bodies as needed.