Rework the mysql connection to use pools
The current server setup has it using just one connection to the database
server. This causes it to crash periodically when the connection to the
database becomes stale. Hopefully using a pool in a manner as Q&D as this
should fix the issue. If not, I'll maybe look into another fix or maybe
there will be just occasional minutes of downtime as PM2 realizes the server
is down and restarts it, IDK.