Export app and underlying web server, update authTester to match
Ensures that we can cleanly close the test server after we have finished with
all of our tests, which was previously impossible. Node would just sit there,
continuing to serve files to the void. This ensures that running `npm test`
will have a chance of completing if you do `server.server.close()` or the like.