Fix return code assertion in delete
The server should only return 400s when there are parameters that could be
missing (i.e. in post or in put) which is not the case in delete. In such a
case, the only possible missing parameter would be the login token, which
would be an authentication error rather than a bad request error.