Clean up the syntax for this a bunch
If we're doing the same tests in all cases (which we essentially are) we can
make this go by a lot smoother by structuring the data, creating a few dummy
function bodies, using forEach() and plugging in where necessary.

This is about ten times easier and I'm so happy that I've found it, honestly.